refactored import code
